added auth and api key generation

+++ src/controllers/BandMember.js
@@ -1,4 +1,6 @@
 const BandMember = require('../models/BandMember');
+const User = require('../models/User');
+const axios = require('axios');
 
 const bandMemberController = {
   getAll: async (req, res) => {
@@ -11,12 +13,15 @@ const bandMemberController = {
   },
   getOne: async (req, res) => {
     try {
-      const bandMember = await BandMember.findById(req.params.id)
-        .populate('albums', 'name releaseDate');
+      const bandMember = await BandMember.findById(req.params.id).populate('albums', 'name releaseDate');
       if (!bandMember) {
         res.status(404).json({ message: 'Band member not found' });
       } else {
-        res.json(bandMember);
+        const API_KEY = process.env.API_KEY;
+        const tenorResponse = await axios.get(`https://tenor.googleapis.com/v2/search?q=${bandMember.name}&limit=1&key=${API_KEY}`);
+        const gifUrl = tenorResponse.data.results[0].media_formats.mediumgif.url;
+        const responseObj = { bandMember: bandMember.toObject(), gifUrl };
+        res.json(responseObj);
       }
     } catch (err) {
       console.log(err);
@@ -25,6 +30,11 @@ const bandMemberController = {
   },
   create: async (req, res) => {
     try {
+      const apiKey = req.query.apiKey;
+      const user = await User.findOne({ apiKey: apiKey });
+      if (!user) {
+        return res.status(401).json({ message: 'Unauthorized' });
+      }
       const newBandMember = await BandMember.create(req.body);
       res.status(201).json(newBandMember);
     } catch (err) {
@@ -33,6 +43,11 @@ const bandMemberController = {
   },
   update: async (req, res) => {
     try {
+      const apiKey = req.query.apiKey;
+      const user = await User.findOne({ apiKey: apiKey });
+      if (!user) {
+        return res.status(401).json({ message: 'Unauthorized' });
+      }
       const bandMember = await BandMember.findById(req.params.id);
       if (!bandMember) {
         res.status(404).json({ message: 'Band member not found' });
@@ -42,22 +57,26 @@ const bandMemberController = {
       }
     } catch (err) {
       res.status(400).json({ message: err.message });
-
     }
   },
   delete: async (req, res) => {
     try {
+      const apiKey = req.query.apiKey;
+      const user = await User.findOne({ apiKey: apiKey });
+      if (!user) {
+        return res.status(401).json({ message: 'Unauthorized' });
+      }
       const bandMember = await BandMember.findById(req.params.id);
       if (!bandMember) {
         res.status(404).json({ message: 'Band member not found' });
+      } else {
+        await bandMember.deleteOne();
+        res.status(200).json({ message: 'Band member deleted' });
       }
-      await bandMember.deleteOne();
-      res.status(200).json({ message: 'Band member deleted' });
     } catch (err) {
       res.status(400).json({ message: err.message });
     }
   }
-
 };
 
 module.exports = bandMemberController;
